passive mode: make external address handed to client configurable
before a data connection in passive mode is opened, the server tells
the client which IP and which PORT to use - this is the responsibility
of the ftpserver and the ftpserver must know which external address
to hand to the client. hence: `passive.externalAddress`

I also added `active.ipCheck` configuration option, since I was already
at it.